Map pin colors correspond with rows in results
When geomapping a cross-search, the color of the map pins now correspond to rows in the search results.  This is done by creating a mapping between lings and the rows that they appear in, and another mapping between colors and the row that they correspond to.

This fix partially remedies the existing problem that the map will stop laying pins when it runs out of colors, since there are typically many lings per row, and often fewer than 20 (max number of colors) rows.  The solution is not scalable, however, and large cross-searches still fail to display all map-pins.